Fixed signing up for stock emails
